data/consumer/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
apache-beam==2.61.0
apache-flink==2.2.0
apache-flink-libraries==2.2.0
attrs==25.4.0
avro==1.12.1
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
dotenv==0.9.9
fastavro==1.12.1
fasteners==0.20
find_libpython==0.5.0
grpcio==1.65.5
h11==0.16.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.1.3
objsize==0.7.1
openai==2.9.0
orjson==3.11.4
packaging==25.0
pandas==2.2.3
pemja==0.5.5
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydantic==2.12.5
pydantic_core==2.41.5
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.15.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zstandard==0.25.0
